requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==4.2.4
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.9.0
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
idna==3.7
oauthlib==3.2.2
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==72.1.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
wheel==0.43.0